fix: 게시글 유형별 조회 좋아요 개수, 댓글 개수, onLike 수정

+++ src/main/java/com/example/everyhanghae/dto/response/BoardDetailResponseDto.java
@@ -18,14 +18,14 @@ public class BoardDetailResponseDto {
     private String content;
     private String userName;
     private boolean onLike;
-    private Long totalLike;
+    private int totalLike;
     private int totalComment;
     private boolean onMine;
     private String createdAt;
 
     private List<CommentResponseDto> commentList;
 
-    public BoardDetailResponseDto(Board board, boolean onLike, Long totalLike, int totalComment, boolean onMine, List<CommentResponseDto> commentResponseDtoList) {
+    public BoardDetailResponseDto(Board board, boolean onLike, int totalLike, int totalComment, boolean onMine, List<CommentResponseDto> commentResponseDtoList) {
         this.id = board.getId();
         this.boardType = board.getBoardType().getBoardType();
         this.typeName = board.getBoardType().getTypeName();
