+++ src/main/java/com/example/everyhanghae/repository/LikesRepository.java
@@ -12,7 +12,7 @@ public interface LikesRepository extends JpaRepository<Likes, Long> {
     Optional<Likes> findByBoardAndUser(Board borard, User user);
 
     //공감 넣기
-    long countByBoard(Board board);
+    int countByBoard(Board board);
 
 //    //게시글 공감 전체 조회
 //    @Query(value = "select count(*) from Likes c where c.board_Id.id = :boardId and c.select = true")
